Add device trees to manifest
recently there seems to be a api/roomservice.py issue when "lunch" command is executed. Roomservice usually takes care of creating "local_manifest" folder then generating a roomservice resource file for the device that is called in the lunch command. This allows repo source to sync the device directory for that said device

This will manually make the device directory to prevent "not found" errors when lunching and executing compilation

P3:
 add Samsung devices

P4:
 add LG & Oppo devices

P5:
 Only Sony left

P6:
 add Sony devices

Signed-off-by: Ayysir <ayysir@aospal.com>
Change-Id: I2cabcfb85c7abcb4f252ac8ad799fbf72d49968f